feat: refactor componentes de certificados y clínicas
