+++ src/app/dashboard/admin/clinics/_components/ClinicsTableActions.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { Edit, MoreHorizontal, RotateCcw, Trash } from "lucide-react";
 
-import { ProtectedComponent } from "@/auth/presentation/components/ProtectedComponent";
 import { Button } from "@/shared/components/ui/button";
 import {
   DropdownMenu,
@@ -13,7 +12,6 @@ import {
 import { useDialogStore } from "@/shared/stores/useDialogStore";
 import { useToggleActiveClinic } from "../_hooks/useClinics";
 import { ClinicResponse } from "../_types/clinics.types";
-import { EnumAction, EnumResource } from "../../roles/_utils/groupedPermission";
 
 interface ClinicsTableActionsProps {
   row: ClinicResponse;
@@ -38,46 +36,35 @@ export default function ClinicsTableActions({ row }: ClinicsTableActionsProps) {
   };
 
   return (
-    <ProtectedComponent
-      requiredPermissions={[
-        { resource: EnumResource.clinics, action: EnumAction.update },
-        { resource: EnumResource.clinics, action: EnumAction.delete },
-      ]}
-    >
-      <DropdownMenu>
-        <DropdownMenuTrigger asChild>
-          <Button variant="ghost" className="bg-background" size="icon">
-            <MoreHorizontal className="size-4" />
-          </Button>
-        </DropdownMenuTrigger>
-        <DropdownMenuContent>
-          <ProtectedComponent requiredPermissions={[{ resource: EnumResource.clinics, action: EnumAction.update }]}>
-            <DropdownMenuItem onClick={handleEdit}>
-              Editar
-              <DropdownMenuShortcut>
-                <Edit className="size-4 mr-2" />
-              </DropdownMenuShortcut>
-            </DropdownMenuItem>
-          </ProtectedComponent>
-          <ProtectedComponent requiredPermissions={[{ resource: EnumResource.clinics, action: EnumAction.delete }]}>
-            {row.isActive ? (
-              <DropdownMenuItem onClick={handleDelete} disabled={!row.isActive}>
-                Eliminar
-                <DropdownMenuShortcut>
-                  <Trash className="size-4 mr-2" />
-                </DropdownMenuShortcut>
-              </DropdownMenuItem>
-            ) : (
-              <DropdownMenuItem onClick={handleToggleActive} disabled={isToggleActivePending}>
-                Reactivar
-                <DropdownMenuShortcut>
-                  <RotateCcw className="size-4 mr-2 text-yellow-600" />
-                </DropdownMenuShortcut>
-              </DropdownMenuItem>
-            )}
-          </ProtectedComponent>
-        </DropdownMenuContent>
-      </DropdownMenu>
-    </ProtectedComponent>
+    <DropdownMenu>
+      <DropdownMenuTrigger asChild>
+        <Button variant="ghost" className="bg-background" size="icon">
+          <MoreHorizontal className="size-4" />
+        </Button>
+      </DropdownMenuTrigger>
+      <DropdownMenuContent>
+        <DropdownMenuItem onClick={handleEdit}>
+          Editar
+          <DropdownMenuShortcut>
+            <Edit className="size-4 mr-2" />
+          </DropdownMenuShortcut>
+        </DropdownMenuItem>
+        {row.isActive ? (
+          <DropdownMenuItem onClick={handleDelete} disabled={!row.isActive}>
+            Eliminar
+            <DropdownMenuShortcut>
+              <Trash className="size-4 mr-2" />
+            </DropdownMenuShortcut>
+          </DropdownMenuItem>
+        ) : (
+          <DropdownMenuItem onClick={handleToggleActive} disabled={isToggleActivePending}>
+            Reactivar
+            <DropdownMenuShortcut>
+              <RotateCcw className="size-4 mr-2 text-yellow-600" />
+            </DropdownMenuShortcut>
+          </DropdownMenuItem>
+        )}
+      </DropdownMenuContent>
+    </DropdownMenu>
   );
 }
