+++ src/app/dashboard/admin/diagnostics/_components/DiagnosticsTableActions.tsx
@@ -2,7 +2,6 @@
 
 import { Edit, FileCheck, FileX, MoreHorizontal, RotateCcw, Trash } from "lucide-react";
 
-import { ProtectedComponent } from "@/auth/presentation/components/ProtectedComponent";
 import { Button } from "@/shared/components/ui/button";
 import {
   DropdownMenu,
@@ -18,7 +17,6 @@ import {
   useToggleIncludeReportsDiagnostic,
 } from "../../medical-records/_hooks/useMedicalRecords";
 import { DiagnosticEntity } from "../../medical-records/_types/medical-record.types";
-import { EnumAction, EnumResource } from "../../roles/_utils/groupedPermission";
 
 interface DiagnosticsTableActionsProps {
   row: DiagnosticEntity;
@@ -50,67 +48,48 @@ export default function DiagnosticsTableActions({ row }: DiagnosticsTableActions
   };
 
   return (
-    <ProtectedComponent
-      requiredPermissions={[
-        { resource: EnumResource.diagnostic, action: EnumAction.update },
-        { resource: EnumResource.diagnostic, action: EnumAction.delete },
-      ]}
-    >
-      <DropdownMenu>
-        <DropdownMenuTrigger asChild>
-          <Button variant="ghost" className="bg-background" size="icon">
-            <MoreHorizontal className="size-4" />
-          </Button>
-        </DropdownMenuTrigger>
-        <DropdownMenuContent>
-          <ProtectedComponent requiredPermissions={[{ resource: EnumResource.diagnostic, action: EnumAction.update }]}>
-            <DropdownMenuItem onClick={handleEdit} disabled={!row.isActive}>
-              Editar
-              <DropdownMenuShortcut>
-                <Edit className="size-4 mr-2" />
-              </DropdownMenuShortcut>
-            </DropdownMenuItem>
-          </ProtectedComponent>
+    <DropdownMenu>
+      <DropdownMenuTrigger asChild>
+        <Button variant="ghost" className="bg-background" size="icon">
+          <MoreHorizontal className="size-4" />
+        </Button>
+      </DropdownMenuTrigger>
+      <DropdownMenuContent>
+        <DropdownMenuItem onClick={handleEdit} disabled={!row.isActive}>
+          Editar
+          <DropdownMenuShortcut>
+            <Edit className="size-4 mr-2" />
+          </DropdownMenuShortcut>
+        </DropdownMenuItem>
 
-          {row.isActive === false && (
-            <ProtectedComponent
-              requiredPermissions={[{ resource: EnumResource.diagnostic, action: EnumAction.update }]}
-            >
-              <DropdownMenuItem onClick={handleActivate} disabled={activateDiagnosticMutation.isPending}>
-                Reactivar
-                <DropdownMenuShortcut>
-                  <RotateCcw className="size-4 mr-2 text-yellow-500" />
-                </DropdownMenuShortcut>
-              </DropdownMenuItem>
-            </ProtectedComponent>
-          )}
-          {row.isActive === true && (
-            <ProtectedComponent
-              requiredPermissions={[{ resource: EnumResource.diagnostic, action: EnumAction.update }]}
-            >
-              <DropdownMenuItem onClick={handleDeactivate} disabled={deactivateDiagnosticMutation.isPending}>
-                Eliminar
-                <DropdownMenuShortcut>
-                  <Trash className="size-4 mr-2 " />
-                </DropdownMenuShortcut>
-              </DropdownMenuItem>
-            </ProtectedComponent>
-          )}
+        {row.isActive === false && (
+          <DropdownMenuItem onClick={handleActivate} disabled={activateDiagnosticMutation.isPending}>
+            Reactivar
+            <DropdownMenuShortcut>
+              <RotateCcw className="size-4 mr-2 text-yellow-500" />
+            </DropdownMenuShortcut>
+          </DropdownMenuItem>
+        )}
+        {row.isActive === true && (
+          <DropdownMenuItem onClick={handleDeactivate} disabled={deactivateDiagnosticMutation.isPending}>
+            Eliminar
+            <DropdownMenuShortcut>
+              <Trash className="size-4 mr-2 " />
+            </DropdownMenuShortcut>
+          </DropdownMenuItem>
+        )}
 
-          <ProtectedComponent requiredPermissions={[{ resource: EnumResource.diagnostic, action: EnumAction.update }]}>
-            <DropdownMenuItem onClick={handleToggleIncludeReports} disabled={toggleIncludeReportsDiagnosticIsPending}>
-              {!row.isDefaultIncluded ? "Incluir en informes" : "Excluir de informes"}
-              <DropdownMenuShortcut>
-                {!row.isDefaultIncluded ? (
-                  <FileCheck className="size-4 mr-2 text-green-500" />
-                ) : (
-                  <FileX className="size-4 mr-2 text-red-500" />
-                )}
-              </DropdownMenuShortcut>
-            </DropdownMenuItem>
-          </ProtectedComponent>
-        </DropdownMenuContent>
-      </DropdownMenu>
-    </ProtectedComponent>
+        <DropdownMenuItem onClick={handleToggleIncludeReports} disabled={toggleIncludeReportsDiagnosticIsPending}>
+          {!row.isDefaultIncluded ? "Incluir en informes" : "Excluir de informes"}
+          <DropdownMenuShortcut>
+            {!row.isDefaultIncluded ? (
+              <FileCheck className="size-4 mr-2 text-green-500" />
+            ) : (
+              <FileX className="size-4 mr-2 text-red-500" />
+            )}
+          </DropdownMenuShortcut>
+        </DropdownMenuItem>
+      </DropdownMenuContent>
+    </DropdownMenu>
   );
 }
